Create a partition object from a data frame

as_partition_step() creates a partition_step object. partition_steps are used while iterating through the partition algorithm: it stores necessary information about how to proceed in the partitioning, such as the information threshold. as_partition_step() is primarily called internally by partition() but can be helpful while developing partitioners.

as_partition_step(
  .x,
  threshold = NA,
  reduced_data = NA,
  target = NA,
  metric = NA,
  tolerance = 0.01,
  var_prefix = NA,
  partitioner = NA,
  ...
)

Arguments

.x
a data.frame or partition_step object

threshold
The minimum information loss allowable

reduced_data
A data set with reduced variables

target
A character or integer vector: the variables to reduce

metric
A measure of information

tolerance
A tolerance around the threshold to accept a reduction

var_prefix
Variable name for reduced variables

partitioner
A partitioner, a part_*() function or one created with as_partitioner().

...
Other objects to store during the partition step

Value

a partition_step object

Examples

.df <- data.frame(x = rnorm(100), y = rnorm(100))
as_partition_step(.df, threshold = .6)
#> $.df
#>               x            y
#> 1   -0.26511667  0.489407998
#> 2    0.10113680 -0.876758691
#> 3    0.33911966  1.916295934
#> 4    0.28163990 -0.536758117
#> 5   -0.73687541  1.643622456
#> 6    0.43203238 -0.602860680
#> 7    0.37832702  0.801702545
#> 8    0.24008038  1.390706762
#> 9    0.71819918  2.758620226
#> 10   0.51145444 -0.770869789
#> 11   0.30958900 -0.027422579
#> 12  -0.77879407 -0.312562106
#> 13  -0.37702279  0.104398706
#> 14  -0.93631719 -0.682961758
#> 15   1.04562010  1.620622388
#> 16   0.80153714 -0.901717207
#> 17   0.78132071 -0.537577582
#> 18   0.40811913  1.191241301
#> 19  -1.17879544  0.249985652
#> 20  -1.51863280 -0.880091594
#> 21   0.44446237 -1.039869849
#> 22   1.05876753  0.059498284
#> 23  -0.14118317 -0.064522463
#> 24  -0.33402703  0.133321848
#> 25  -0.73304914 -1.093606506
#> 26   1.37941182  1.277416777
#> 27   0.20107973 -0.870821454
#> 28  -0.49936594  0.701231381
#> 29   1.39181035  0.008273018
#> 30  -0.62142978 -0.631331615
#> 31   2.22553572  0.247537818
#> 32  -0.43670206  1.263425821
#> 33  -0.17918512  0.613112120
#> 34   1.88362696 -0.836531139
#> 35  -0.13185665 -0.840099365
#> 36   0.45694710  0.449535715
#> 37   0.34294324  0.630851881
#> 38   0.52579393  2.183186284
#> 39  -2.04557859 -1.082647155
#> 40  -0.15083636  1.141617779
#> 41  -0.56193327 -1.207157224
#> 42   1.07662366  0.139269552
#> 43  -0.53175578  0.383397328
#> 44   1.68315334  0.276986676
#> 45  -0.13500245  0.802912723
#> 46  -1.00747289  2.281559576
#> 47  -1.88422852  1.621321661
#> 48   0.04807336 -0.387563702
#> 49  -0.53106503 -0.917862471
#> 50   1.05776282 -0.426105236
#> 51  -0.83485202  0.491754475
#> 52   1.07992803 -1.502152183
#> 53  -0.06087972  0.041806696
#> 54  -0.57199857 -2.108681826
#> 55   1.26114308 -0.556901141
#> 56   0.60873165  0.535000677
#> 57  -2.21235745  0.953000582
#> 58  -0.72910683 -0.107675576
#> 59  -0.79618173 -0.610504712
#> 60   0.87216022 -0.286148897
#> 61  -0.39989309 -0.793279068
#> 62   1.48339790  1.009940642
#> 63   1.01586754  0.196495298
#> 64   1.17797292  1.963664497
#> 65  -0.87481791 -0.505748951
#> 66   0.94012892 -1.059759028
#> 67   0.90881597 -0.472749965
#> 68   1.32248867 -0.752478261
#> 69   0.44271775 -1.116270429
#> 70   0.17992826 -0.249390343
#> 71   0.11651097 -1.311975766
#> 72   1.80431990  0.320953248
#> 73   0.16557529  0.481862419
#> 74  -1.74423632  0.589984606
#> 75   0.60043119  0.045237319
#> 76  -0.50004311  0.790034744
#> 77  -0.97084395 -0.717404951
#> 78  -1.46675416  0.110816413
#> 79   0.02218023 -0.666045171
#> 80   0.20357082 -0.365675085
#> 81  -0.74035560 -0.689284076
#> 82  -1.52879822  1.226595405
#> 83   0.84225868  0.486472658
#> 84   0.72943444 -0.549058720
#> 85  -1.45746425  0.473904816
#> 86  -1.62594250  0.783590958
#> 87   1.79367398  0.430071766
#> 88  -1.47346668 -0.302234965
#> 89  -0.90871609  1.483885188
#> 90   0.13047307  0.347635477
#> 91  -0.50717060  1.002412555
#> 92   0.38995475  0.113247821
#> 93   0.65349272  0.284849808
#> 94   1.22574181 -0.820137162
#> 95   0.45389025  0.093334291
#> 96  -1.39010633  0.616152863
#> 97   0.95879850  1.128748497
#> 98  -0.38839312 -0.456289814
#> 99  -1.84776501 -0.002214323
#> 100 -0.11523702 -0.733182978
#> 
#> $threshold
#> [1] 0.6
#> 
#> $target
#> [1] NA
#> 
#> $last_target
#> [1] NA
#> 
#> $reduced_data
#> # A tibble: 100 × 2
#>         x      y
#>     <dbl>  <dbl>
#>  1 -0.265  0.489
#>  2  0.101 -0.877
#>  3  0.339  1.92 
#>  4  0.282 -0.537
#>  5 -0.737  1.64 
#>  6  0.432 -0.603
#>  7  0.378  0.802
#>  8  0.240  1.39 
#>  9  0.718  2.76 
#> 10  0.511 -0.771
#> # … with 90 more rows
#> 
#> $metric
#> [1] NA
#> 
#> $tolerance
#> [1] 0.01
#> 
#> $mapping_key
#> # A tibble: 2 × 3
#>   variable mapping   information
#>   <chr>    <list>          <dbl>
#> 1 x        <chr [1]>           1
#> 2 y        <chr [1]>           1
#> 
#> $var_prefix
#> [1] NA
#> 
#> $all_done
#> [1] FALSE
#> 
#> $partitioner
#> [1] NA
#> 
#> attr(,"class")
#> [1] "partition_step"