Measure the information loss of reduction using the minimum intraclass correlation coefficient

Metrics are functions that tell how much information would be lost for a given reduction in the data. reduce. as_measure() is a helper function to create new metrics to be used in partitioners. partitioners can be created with as_partitioner().

measure_min_icc() assesses information loss by calculating the intraclass correlation coefficient for each set of the target variables and finding their minimum.

measure_min_icc(.partition_step, search_method = c("binary", "linear"))

Arguments

.partition_step

a partition_step object

search_method

The search method. Binary search is generally more efficient but linear search can be faster in very low dimensions.

Value

a partition_step object

See also