Partitioner: distance, mutual information, scaled means

Partitioners are functions that tell the partition algorithm 1) what to try to reduce 2) how to measure how much information is lost from the reduction and 3) how to reduce the data. In partition, functions that handle 1) are called directors, functions that handle 2) are called metrics, and functions that handle 3) are called reducers. partition has a number of pre-specified partitioners for agglomerative data reduction. Custom partitioners can be created with as_partitioner().

Pass partitioner objects to the partitioner argument of partition().

part_stdmi() uses the following direct-measure-reduce approach:

direct: direct_distance(), Minimum Distance

measure: measure_std_mutualinfo(), Standardized Mutual Information

reduce: reduce_scaled_mean(), Scaled Row Means

part_stdmi(spearman = FALSE)

Arguments

spearman

logical. Use Spearman's correlation for distance matrix?

Value

a partitioner

See also

Examples

set.seed(123)
df <- simulate_block_data(c(3, 4, 5), lower_corr = .4, upper_corr = .6, n = 100)

# fit partition using part_stdmi()
partition(df, threshold = .6, partitioner = part_stdmi())
#> Partitioner:
#>    Director: Minimum Distance (Pearson) 
#>    Metric: Standardized Mutual Information 
#>    Reducer: Scaled Mean
#> 
#> Reduced Variables:
#> 0 reduced variables created from 0 observed variables