Simulate correlated blocks of variables

simulate_block_data() creates a dataset of blocks of data where variables within each block are correlated. The correlation for each pair of variables is sampled uniformly from lower_corr to upper_corr, and the values of each are sampled using MASS::mvrnorm().

simulate_block_data(
  block_sizes,
  lower_corr,
  upper_corr,
  n,
  block_name = "block",
  sep = "_",
  var_name = "x"
)

Arguments

block_sizes
a vector of block sizes. The size of each block is the number of variables within it.

lower_corr
the lower bound of the correlation within each block

upper_corr
the upper bound of the correlation within each block

n
the number of observations or rows

block_name
description prepended to the variable to indicate the block it belongs to

sep
a character, what to separate the variable names with

var_name
the name of the variable within the block

Value

a tibble with sum(block_sizes) columns and n rows.

Examples

# create a 100 x 15 data set with 3 blocks
simulate_block_data(
  block_sizes = rep(5, 3),
  lower_corr = .4,
  upper_corr = .6,
  n = 100
)

#> # A tibble: 100 x 15
#>    block1_x1 block1_x2 block1_x3 block1_x4 block1_x5 block2_x1 block2_x2
#>        <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>
#>  1    0.866     0.0101     1.57     0.535      0.652   -0.874     0.0717
#>  2    0.149     1.02       0.202   -0.559      0.926   -1.10      1.27  
#>  3   -0.221     1.20       0.696    0.460     -0.212    1.19      1.92  
#>  4    1.29      0.447      1.08    -0.172     -0.159   -0.667    -0.0464
#>  5    0.938     1.47       1.11    -0.129      0.557   -1.26     -1.07  
#>  6   -0.0644   -1.39       1.15     0.0467    -0.962   -1.70     -0.633 
#>  7    0.281    -0.0811     0.915   -1.68      -1.55    -0.955    -0.134 
#>  8   -0.274     1.19       0.569   -0.287      1.19     0.274     0.593 
#>  9   -1.47     -1.22      -0.287   -0.460     -1.77     0.510    -0.729 
#> 10   -0.0332   -1.76       0.841   -1.74      -1.22    -0.0389    1.57  
#> # … with 90 more rows, and 8 more variables: block2_x3 <dbl>, block2_x4 <dbl>,
#> #   block2_x5 <dbl>, block3_x1 <dbl>, block3_x2 <dbl>, block3_x3 <dbl>,
#> #   block3_x4 <dbl>, block3_x5 <dbl>