Create a partition object from a data frame

as_partition_step() creates a partition_step object. partition_steps are used while iterating through the partition algorithm: it stores necessary information about how to proceed in the partitioning, such as the information threshold. as_partition_step() is primarily called internally by partition() but can be helpful while developing partitioners.

as_partition_step(
  .x,
  threshold = NA,
  reduced_data = NA,
  target = NA,
  metric = NA,
  tolerance = 0.01,
  var_prefix = NA,
  partitioner = NA,
  ...
)

Arguments

.x
a data.frame or partition_step object

threshold
The minimum information loss allowable

reduced_data
A data set with reduced variables

target
A character or integer vector: the variables to reduce

metric
A measure of information

tolerance
A tolerance around the threshold to accept a reduction

var_prefix
Variable name for reduced variables

partitioner
A partitioner, a part_*() function or one created with as_partitioner().

...
Other objects to store during the partition step

Value

a partition_step object

Examples

.df <- data.frame(x = rnorm(100), y = rnorm(100))
as_partition_step(.df, threshold = .6)
#> $.df
#>               x            y
#> 1   -0.23579128 -0.246326437
#> 2   -0.12072917  0.852473368
#> 3    0.99379952 -0.032401873
#> 4    0.20102409  0.353233550
#> 5    0.24358353  0.262414712
#> 6    0.67561130  0.755273123
#> 7    0.24594943 -2.378990858
#> 8   -1.08214641 -1.231623280
#> 9   -0.38488212  1.212019487
#> 10   0.65464270  0.758969843
#> 11   0.13776009 -0.177511123
#> 12   0.25191121  1.055588703
#> 13  -0.79171673  0.547514788
#> 14   1.44686632 -0.001215662
#> 15  -0.65983510 -0.078884087
#> 16  -0.76998215 -0.502603795
#> 17   0.16233783  0.670317805
#> 18  -1.21942251 -0.122328782
#> 19   0.48563487  0.462543474
#> 20   1.51704086 -0.171504098
#> 21  -0.12321437 -0.433892392
#> 22   0.93560895 -0.989277440
#> 23  -0.67478093  0.118108516
#> 24  -0.77617806  0.122711669
#> 25   0.11516689 -0.652972740
#> 26   0.27689564  1.873881566
#> 27   0.81598296  1.171038431
#> 28   1.23660452 -0.368165082
#> 29  -0.34891737  0.478383678
#> 30  -0.29121039  0.774358772
#> 31   1.87175867  0.659714538
#> 32   1.15581821 -1.652526444
#> 33  -0.88955792  0.427601874
#> 34   1.17587588  0.081505533
#> 35  -0.95463354 -0.687440488
#> 36   0.02523381  1.231709142
#> 37   0.67495674  0.284495828
#> 38   0.71991128 -0.072611665
#> 39   0.93650482 -0.971845027
#> 40   1.07110140  0.790791295
#> 41   0.51473670  2.196321379
#> 42  -0.66242760  1.910064591
#> 43  -0.13758662 -0.195744501
#> 44   0.78000991 -1.029709346
#> 45  -1.46866995  0.757690047
#> 46   0.53332668 -0.436707642
#> 47  -1.46146949 -0.387063455
#> 48  -1.00228525 -0.776784239
#> 49   1.27465436  0.308725510
#> 50  -1.03840349 -0.375803403
#> 51   1.29018416 -0.326772429
#> 52  -0.46507981 -1.134978688
#> 53   2.67393572  2.223278759
#> 54  -0.33142513  0.393561460
#> 55  -0.67983019 -1.820605472
#> 56   0.09967796 -1.663356114
#> 57  -2.31889668  0.926530215
#> 58   0.23118018  0.495190878
#> 59   0.29003774  0.459689472
#> 60   0.08955173 -0.478012414
#> 61  -0.24110201  1.148009342
#> 62   0.58641883 -0.424456217
#> 63   1.02264408 -2.212509904
#> 64   0.20052794  0.481966639
#> 65  -0.92854089 -1.186617272
#> 66  -1.20425563  1.286832608
#> 67  -1.83519263  0.970975747
#> 68  -0.94515948  0.679763211
#> 69   0.04133604  0.243438167
#> 70   0.59599213 -0.257565515
#> 71  -0.28133840 -0.754059326
#> 72  -0.48892218 -0.733006661
#> 73  -0.09374387  0.836966511
#> 74   0.09117453  0.096684990
#> 75   1.03471549  0.235945163
#> 76  -0.49344878  0.359105527
#> 77   0.10562851  0.825590913
#> 78  -0.06939424 -1.082306747
#> 79  -1.66665101  0.222156824
#> 80  -0.10329282  0.057515780
#> 81   1.53100395  1.562312766
#> 82   1.25384733 -0.074216003
#> 83   1.26413255  0.862966537
#> 84   1.20552473  1.082737237
#> 85  -0.59740424 -0.301897760
#> 86  -0.79556337 -0.992293023
#> 87  -0.68060624  0.029548189
#> 88  -1.06830542 -1.017935919
#> 89   0.56252916  0.680214419
#> 90  -0.43007449 -0.180759962
#> 91  -0.03760148 -0.156311431
#> 92  -0.85407460 -1.949806176
#> 93  -0.22967182 -0.210862806
#> 94  -1.42576704  0.135372423
#> 95   0.84989385  0.130242228
#> 96   1.67624981 -0.085476355
#> 97  -1.10065643 -0.040653903
#> 98  -0.78512052 -0.631877784
#> 99   0.39212810 -0.246841492
#> 100  0.46113522  0.206586526
#> 
#> $threshold
#> [1] 0.6
#> 
#> $target
#> [1] NA
#> 
#> $last_target
#> [1] NA
#> 
#> $reduced_data
#> # A tibble: 100 × 2
#>         x       y
#>     <dbl>   <dbl>
#>  1 -0.236 -0.246 
#>  2 -0.121  0.852 
#>  3  0.994 -0.0324
#>  4  0.201  0.353 
#>  5  0.244  0.262 
#>  6  0.676  0.755 
#>  7  0.246 -2.38  
#>  8 -1.08  -1.23  
#>  9 -0.385  1.21  
#> 10  0.655  0.759 
#> # … with 90 more rows
#> 
#> $metric
#> [1] NA
#> 
#> $tolerance
#> [1] 0.01
#> 
#> $mapping_key
#> # A tibble: 2 × 3
#>   variable mapping   information
#>   <chr>    <list>          <dbl>
#> 1 x        <chr [1]>           1
#> 2 y        <chr [1]>           1
#> 
#> $var_prefix
#> [1] NA
#> 
#> $all_done
#> [1] FALSE
#> 
#> $partitioner
#> [1] NA
#> 
#> attr(,"class")
#> [1] "partition_step"