Create a partition object from a data frame

as_partition_step() creates a partition_step object. partition_steps are used while iterating through the partition algorithm: it stores necessary information about how to proceed in the partitioning, such as the information threshold. as_partition_step() is primarily called internally by partition() but can be helpful while developing partitioners.

as_partition_step(
  .x,
  threshold = NA,
  reduced_data = NA,
  target = NA,
  metric = NA,
  tolerance = 0.01,
  var_prefix = NA,
  partitioner = NA,
  ...
)

Arguments

.x
a data.frame or partition_step object

threshold
The minimum information loss allowable

reduced_data
A data set with reduced variables

target
A character or integer vector: the variables to reduce

metric
A measure of information

tolerance
A tolerance around the threshold to accept a reduction

var_prefix
Variable name for reduced variables

partitioner
A partitioner, a part_*() function or one created with as_partitioner().

...
Other objects to store during the partition step

Value

a partition_step object

Examples

.df <- data.frame(x = rnorm(100), y = rnorm(100))
as_partition_step(.df, threshold = .6)
#> $.df
#>               x            y
#> 1   -1.12228940  0.780356304
#> 2    1.61502817 -0.465940250
#> 3   -1.60007293 -0.744476080
#> 4    0.51376497  0.316539029
#> 5    1.82241221  0.124283573
#> 6    0.45610536  0.081069277
#> 7    2.23445513  1.736339522
#> 8    1.42294777  0.980518048
#> 9   -0.32611510 -0.296786748
#> 10  -1.23885143  1.016862081
#> 11   2.36608912 -0.689523897
#> 12  -0.80954910 -0.384600745
#> 13  -0.08066577 -0.700700448
#> 14   1.48160528 -0.920894083
#> 15  -1.14875245 -2.663507994
#> 16   0.53228818  1.353093091
#> 17   2.01010611  0.710872505
#> 18   0.66126365  0.652924870
#> 19   1.32192284  0.382302001
#> 20   0.11262487  1.012266283
#> 21  -1.12677439 -0.007300406
#> 22   0.04140505  0.969061935
#> 23   0.12680814 -0.535643048
#> 24  -0.06605754 -1.063721256
#> 25   0.37200676 -1.566005470
#> 26  -1.57564639  0.381760089
#> 27  -0.54464652 -2.156526564
#> 28   1.26898591 -0.300756757
#> 29   0.21872850 -0.951336653
#> 30  -0.10493179  0.762056172
#> 31   1.75876612 -1.081311095
#> 32  -1.71183614 -1.807292103
#> 33   0.62842789  1.206167910
#> 34  -1.21341488 -0.485172441
#> 35   0.05660853 -0.402335715
#> 36  -1.09185888 -1.028091774
#> 37   0.86322289  0.608459554
#> 38   0.10820733 -0.878650729
#> 39  -0.12571387  1.542847205
#> 40   1.40497454 -1.367500282
#> 41   0.03798754  0.318042108
#> 42   0.32312875 -0.085671684
#> 43  -0.74587917  0.106458468
#> 44  -1.55792081  0.623486917
#> 45   0.63667845  1.621223028
#> 46   1.50508334 -1.080550440
#> 47  -0.25521557 -0.963317701
#> 48   1.60120093 -0.388415499
#> 49  -0.24340879 -1.725855321
#> 50   0.97806283  0.113968349
#> 51  -1.55662767 -0.782379629
#> 52   1.03080369  1.066428071
#> 53   0.47329183 -0.625667949
#> 54   1.33261039 -0.089672367
#> 55   0.47170891  0.583555842
#> 56   0.05849490 -0.494010961
#> 57   1.62036486 -0.690677369
#> 58   0.83418722 -0.472538255
#> 59   0.42422318  0.074030287
#> 60   0.16770674  1.226945800
#> 61  -1.17372184  0.619556632
#> 62  -2.09540262 -1.568074141
#> 63   0.28477570 -0.084033665
#> 64   0.74992339 -0.750103894
#> 65   1.17105644  2.318640111
#> 66   1.74200227 -2.807032880
#> 67   1.16228540 -0.036023224
#> 68   2.17695756 -0.298453707
#> 69  -0.75130625 -0.447166215
#> 70  -0.06592487 -0.848470075
#> 71   0.86419094  1.550305243
#> 72   0.03638611 -2.055642340
#> 73  -0.23617878  0.358806681
#> 74   0.19652356 -0.822803168
#> 75   0.54521182 -1.292437933
#> 76  -0.74886775 -0.654366503
#> 77   1.00054563  1.219244345
#> 78   1.03490991  1.254650874
#> 79  -1.34993421 -1.063736692
#> 80   0.36168448 -2.788051965
#> 81  -1.01817603  0.900181402
#> 82   0.03222117 -0.589183914
#> 83   0.50389248  0.002735699
#> 84   0.62544427  0.636719410
#> 85  -1.40793828 -1.627539315
#> 86   0.03739924  0.071940173
#> 87   1.11706386  1.625134583
#> 88  -1.33725151 -1.465001536
#> 89  -0.75483578  0.178771450
#> 90  -2.04875274  0.753212205
#> 91  -0.52274741 -0.334096054
#> 92  -1.85908648 -0.310073184
#> 93   0.05851254  0.874382487
#> 94  -1.42620557  0.587573219
#> 95   0.36071392  0.769575257
#> 96   0.07309595  0.497207424
#> 97   0.39473903  0.559282366
#> 98   0.41951736  1.032421413
#> 99  -0.12671413  0.728431725
#> 100 -0.43664861 -0.387604614
#> 
#> $threshold
#> [1] 0.6
#> 
#> $target
#> [1] NA
#> 
#> $last_target
#> [1] NA
#> 
#> $reduced_data
#> # A tibble: 100 × 2
#>         x       y
#>     <dbl>   <dbl>
#>  1 -1.12   0.780 
#>  2  1.62  -0.466 
#>  3 -1.60  -0.744 
#>  4  0.514  0.317 
#>  5  1.82   0.124 
#>  6  0.456  0.0811
#>  7  2.23   1.74  
#>  8  1.42   0.981 
#>  9 -0.326 -0.297 
#> 10 -1.24   1.02  
#> # … with 90 more rows
#> 
#> $metric
#> [1] NA
#> 
#> $tolerance
#> [1] 0.01
#> 
#> $mapping_key
#> # A tibble: 2 × 3
#>   variable mapping   information
#>   <chr>    <list>          <dbl>
#> 1 x        <chr [1]>           1
#> 2 y        <chr [1]>           1
#> 
#> $var_prefix
#> [1] NA
#> 
#> $all_done
#> [1] FALSE
#> 
#> $partitioner
#> [1] NA
#> 
#> attr(,"class")
#> [1] "partition_step"