Create a partition object from a data frame

as_partition_step() creates a partition_step object. partition_steps are used while iterating through the partition algorithm: it stores necessary information about how to proceed in the partitioning, such as the information threshold. as_partition_step() is primarily called internally by partition() but can be helpful while developing partitioners.

as_partition_step(
  .x,
  threshold = NA,
  reduced_data = NA,
  target = NA,
  metric = NA,
  tolerance = 0.01,
  var_prefix = NA,
  partitioner = NA,
  ...
)

Arguments

.x
a data.frame or partition_step object

threshold
The minimum information loss allowable

reduced_data
A data set with reduced variables

target
A character or integer vector: the variables to reduce

metric
A measure of information

tolerance
A tolerance around the threshold to accept a reduction

var_prefix
Variable name for reduced variables

partitioner
A partitioner, a part_*() function or one created with as_partitioner().

...
Other objects to store during the partition step

Value

a partition_step object

Examples

.df <- data.frame(x = rnorm(100), y = rnorm(100))
as_partition_step(.df, threshold = .6)
#> $.df
#>               x           y
#> 1    0.35929034 -0.36764376
#> 2   -0.42134385 -0.91026145
#> 3    0.45765601 -0.56996231
#> 4   -1.14938572  0.93235903
#> 5   -0.32016266 -0.36404706
#> 6   -0.60858468 -1.30742035
#> 7   -1.26397015 -1.58997455
#> 8   -0.32091051 -0.80926262
#> 9    0.24923832 -0.91189574
#> 10  -0.36829219  1.93676609
#> 11   2.03374483  0.62176069
#> 12   1.25872770 -0.92356673
#> 13  -0.30500405  0.05621364
#> 14  -1.06002324  0.40718061
#> 15  -0.80501892 -0.27316537
#> 16  -0.69314197  1.08364126
#> 17   1.18341257  0.09296387
#> 18  -0.08956049 -0.16844115
#> 19  -1.05946579 -0.73163122
#> 20   1.69088093 -1.67953049
#> 21  -0.15150941 -0.53416194
#> 22  -0.34377905 -0.44735378
#> 23   2.69888758  0.11186877
#> 24  -1.18605098 -1.12576684
#> 25  -1.08674004 -0.10918876
#> 26   1.20382991  0.51192405
#> 27   1.62526308  0.20530430
#> 28  -0.24903818 -1.20709516
#> 29  -1.21068136 -0.04941792
#> 30  -1.45978190 -0.61058516
#> 31  -0.17518889  1.29533471
#> 32   0.32522315  0.59978454
#> 33  -0.50383365 -2.00984019
#> 34  -1.45411119 -1.00934298
#> 35  -2.23769033 -1.07376610
#> 36   0.18690379 -0.72889417
#> 37   0.74524016  0.13859293
#> 38   2.12206708  0.01036295
#> 39   0.48829221  1.42576362
#> 40  -0.14388944 -0.28686276
#> 41  -0.37542633  0.45778004
#> 42  -0.05383800  1.38139194
#> 43  -0.93615424 -0.37260071
#> 44  -1.46926733 -0.08658574
#> 45  -0.86775121 -0.73116777
#> 46   0.76108201  0.01959471
#> 47   1.63845670 -2.21910392
#> 48   1.98461631  0.01096067
#> 49  -0.62330516  0.20784011
#> 50  -0.92218931  0.52040106
#> 51   1.03856569  0.89974133
#> 52   1.61617274  0.55587692
#> 53   0.30167811 -0.04609538
#> 54   0.60063649  0.16220820
#> 55  -2.22418363  0.10083569
#> 56   1.07951192 -1.58110079
#> 57   0.07746768  0.63338301
#> 58   1.80239839  0.05993284
#> 59   0.04379777 -2.23670940
#> 60   1.39994607 -1.21652715
#> 61  -0.37612353 -0.13979341
#> 62  -1.52130461  0.78779476
#> 63  -0.61258315  0.07501639
#> 64  -0.04910176  0.50187264
#> 65  -1.15168966  0.33866304
#> 66  -0.37306858  0.52916942
#> 67   0.72864429  0.80673656
#> 68   0.78663973 -0.46481841
#> 69   1.33955933  0.62534549
#> 70   0.27107437 -0.74593230
#> 71   0.41183996  1.59893393
#> 72   1.82050175 -0.55732698
#> 73   0.87490923  0.68291436
#> 74   0.18813855  1.71481577
#> 75  -0.82313631 -0.90882378
#> 76  -0.06985143 -0.31162627
#> 77  -1.35549368  1.39364070
#> 78   0.26756222  1.44640495
#> 79   0.32871071 -2.03126886
#> 80  -0.28019868 -0.67968004
#> 81   1.37268811  0.28782589
#> 82   0.22585280 -0.74449252
#> 83  -0.46777392  1.00909014
#> 84   0.80673216 -1.13557337
#> 85  -0.98637042  1.15496447
#> 86   0.60325887  0.55134519
#> 87   0.01899437 -0.45290100
#> 88  -0.34995824 -0.60484228
#> 89   0.72064745  0.05921967
#> 90  -0.72779357  0.67996201
#> 91  -0.17261205 -0.65549326
#> 92  -0.32778936 -2.11895531
#> 93  -0.60449299 -1.24530179
#> 94  -1.15692186  1.03910596
#> 95  -0.23262336  1.07505332
#> 96   0.27108886  1.11258491
#> 97   0.74642771  0.29396016
#> 98  -1.20003696  0.19683573
#> 99  -0.79689607  1.08348721
#> 100 -0.75760755  1.62453938
#> 
#> $threshold
#> [1] 0.6
#> 
#> $target
#> [1] NA
#> 
#> $last_target
#> [1] NA
#> 
#> $reduced_data
#> # A tibble: 100 × 2
#>         x      y
#>     <dbl>  <dbl>
#>  1  0.359 -0.368
#>  2 -0.421 -0.910
#>  3  0.458 -0.570
#>  4 -1.15   0.932
#>  5 -0.320 -0.364
#>  6 -0.609 -1.31 
#>  7 -1.26  -1.59 
#>  8 -0.321 -0.809
#>  9  0.249 -0.912
#> 10 -0.368  1.94 
#> # … with 90 more rows
#> 
#> $metric
#> [1] NA
#> 
#> $tolerance
#> [1] 0.01
#> 
#> $mapping_key
#> # A tibble: 2 × 3
#>   variable mapping   information
#>   <chr>    <list>          <dbl>
#> 1 x        <chr [1]>           1
#> 2 y        <chr [1]>           1
#> 
#> $var_prefix
#> [1] NA
#> 
#> $all_done
#> [1] FALSE
#> 
#> $partitioner
#> [1] NA
#> 
#> attr(,"class")
#> [1] "partition_step"