Create a partition object from a data frame

as_partition_step() creates a partition_step object. partition_steps are used while iterating through the partition algorithm: it stores necessary information about how to proceed in the partitioning, such as the information threshold. as_partition_step() is primarily called internally by partition() but can be helpful while developing partitioners.

as_partition_step(
  .x,
  threshold = NA,
  reduced_data = NA,
  target = NA,
  metric = NA,
  tolerance = 0.01,
  var_prefix = NA,
  partitioner = NA,
  ...
)

Arguments

.x
a data.frame or partition_step object

threshold
The minimum information loss allowable

reduced_data
A data set with reduced variables

target
A character or integer vector: the variables to reduce

metric
A measure of information

tolerance
A tolerance around the threshold to accept a reduction

var_prefix
Variable name for reduced variables

partitioner
A partitioner, a part_*() function or one created with as_partitioner().

...
Other objects to store during the partition step

Value

a partition_step object

Examples

.df <- data.frame(x = rnorm(100), y = rnorm(100))
as_partition_step(.df, threshold = .6)
#> $.df
#>                x           y
#> 1    0.881672860  0.31684982
#> 2    0.494912498  0.71407626
#> 3   -0.172629276  0.88513610
#> 4    0.592719505  0.15349287
#> 5    0.253225127 -0.27354773
#> 6    0.805872292 -1.57396440
#> 7    0.388594004  0.76980691
#> 8    0.118389978  0.03056990
#> 9   -0.351903128 -0.58494626
#> 10  -1.196182336 -0.06433438
#> 11   0.910514809 -0.71402372
#> 12  -0.692165794  0.31779051
#> 13   0.132263532 -0.32144732
#> 14  -0.304153836  0.08912897
#> 15   0.504505423  0.72444654
#> 16  -0.476809694 -0.17972727
#> 17   2.362731301  0.39240150
#> 18  -0.474946099 -0.22896030
#> 19  -0.418252917 -1.14314683
#> 20   0.134667413  0.59032742
#> 21   1.247047764  1.12247212
#> 22  -0.457895192  0.60175374
#> 23  -0.037319089 -0.31054248
#> 24   0.870018519  0.62163531
#> 25   0.599291085 -0.40265264
#> 26   0.527408726 -0.34313388
#> 27   1.003887123  1.65649369
#> 28   0.796613908  0.41074746
#> 29   0.281196122  0.13825475
#> 30  -0.566144698 -0.33303764
#> 31  -0.303512080  1.33498132
#> 32  -0.638298399  1.34309888
#> 33  -0.842158803 -0.63077472
#> 34   0.292766583 -0.69580839
#> 35  -0.518289563  1.63672984
#> 36   1.142588853  0.48898505
#> 37   1.780330773  1.05849239
#> 38  -0.356455583 -0.50958490
#> 39   0.152667458 -1.24662667
#> 40  -0.103583060  1.08620613
#> 41  -2.075302169  1.39506874
#> 42  -0.948592614 -0.64935268
#> 43  -0.019335775 -1.32517644
#> 44  -0.644260419  0.55412842
#> 45  -0.501452011 -1.15252307
#> 46   1.184722486  0.89405133
#> 47  -2.591939123 -1.81797781
#> 48  -1.459860813  0.46181675
#> 49  -0.009810846 -0.00953944
#> 50  -0.731346209 -0.06483269
#> 51   0.392750921 -0.07584184
#> 52   1.366692840  2.76812720
#> 53   0.895663337  0.43106942
#> 54  -0.061963748  0.59495269
#> 55  -1.087372451 -0.68278577
#> 56  -0.567682438  0.28414238
#> 57   0.456832911 -1.18670174
#> 58  -0.691056995 -1.72200410
#> 59  -0.866279792 -0.57154470
#> 60   0.042107505  0.81206826
#> 61   0.307926429 -0.18582813
#> 62   0.365670433  2.95086529
#> 63  -1.795411625 -0.79444499
#> 64   0.167457106 -0.37853406
#> 65  -1.875616344 -0.04227925
#> 66  -0.374829373 -0.93047677
#> 67  -1.382588478  0.98978105
#> 68   1.207676161  0.24134777
#> 69   2.459137442 -0.83303828
#> 70  -0.713627337 -0.37014983
#> 71  -0.130137904  0.11818254
#> 72   0.865473617  0.14050785
#> 73  -1.243893019 -1.12355746
#> 74   0.159566158  0.32077448
#> 75   0.206268138 -2.20822372
#> 76   0.685978964  1.00160363
#> 77  -0.021365985  1.70383113
#> 78   0.314660293  0.49045788
#> 79   1.099944839  0.07497098
#> 80   0.657150139 -1.51171893
#> 81   1.501397017 -3.08175942
#> 82   0.148885276  0.43631201
#> 83  -0.053830172  0.46512006
#> 84  -0.893521981  0.12739690
#> 85   1.283207174 -1.25762705
#> 86   1.843656465  1.24791841
#> 87   0.833615927 -1.18177188
#> 88  -0.630970372  0.86026138
#> 89   1.022704787 -1.94419427
#> 90   2.346325484 -0.13009774
#> 91   0.012973589  0.20679756
#> 92   0.181363184 -0.42734488
#> 93   0.222655944  0.20034124
#> 94   0.255698716  0.42005474
#> 95  -0.247818459  1.55087578
#> 96  -0.154076830  0.16393995
#> 97   0.080771059  0.82958834
#> 98  -1.539831151 -0.79742307
#> 99   0.496894255 -0.31862573
#> 100 -0.344309268 -0.35285218
#> 
#> $threshold
#> [1] 0.6
#> 
#> $target
#> [1] NA
#> 
#> $last_target
#> [1] NA
#> 
#> $reduced_data
#> # A tibble: 100 × 2
#>         x       y
#>     <dbl>   <dbl>
#>  1  0.882  0.317 
#>  2  0.495  0.714 
#>  3 -0.173  0.885 
#>  4  0.593  0.153 
#>  5  0.253 -0.274 
#>  6  0.806 -1.57  
#>  7  0.389  0.770 
#>  8  0.118  0.0306
#>  9 -0.352 -0.585 
#> 10 -1.20  -0.0643
#> # … with 90 more rows
#> 
#> $metric
#> [1] NA
#> 
#> $tolerance
#> [1] 0.01
#> 
#> $mapping_key
#> # A tibble: 2 × 3
#>   variable mapping   information
#>   <chr>    <list>          <dbl>
#> 1 x        <chr [1]>           1
#> 2 y        <chr [1]>           1
#> 
#> $var_prefix
#> [1] NA
#> 
#> $all_done
#> [1] FALSE
#> 
#> $partitioner
#> [1] NA
#> 
#> attr(,"class")
#> [1] "partition_step"