Replace the director, metric, or reducer for a partitioner

Replace the director, metric, or reducer for a partitioner

replace_partitioner(partitioner, direct = NULL, measure = NULL, reduce = NULL)

Arguments

partitioner

a partitioner

direct

a function that directs, possibly created by as_director()

measure

a function that measures, possibly created by as_measure()

reduce

a function that reduces, possibly created by as_reducer()

Value

a partitioner

See also

Examples

replace_partitioner(
  part_icc,
  reduce = as_reducer(rowMeans)
)
#>    Director: Minimum Distance (Pearson) 
#>    Metric: Intraclass Correlation 
#>    Reducer: <custom reducer>