Simulate correlated blocks of variables

simulate_block_data() creates a dataset of blocks of data where variables within each block are correlated. The correlation for each pair of variables is sampled uniformly from lower_corr to upper_corr, and the values of each are sampled using MASS::mvrnorm().

simulate_block_data(
  block_sizes,
  lower_corr,
  upper_corr,
  n,
  block_name = "block",
  sep = "_",
  var_name = "x"
)

Arguments

block_sizes
a vector of block sizes. The size of each block is the number of variables within it.

lower_corr
the lower bound of the correlation within each block

upper_corr
the upper bound of the correlation within each block

n
the number of observations or rows

block_name
description prepended to the variable to indicate the block it belongs to

sep
a character, what to separate the variable names with

var_name
the name of the variable within the block

Value

a tibble with sum(block_sizes) columns and n rows.

Examples

# create a 100 x 15 data set with 3 blocks
simulate_block_data(
  block_sizes = rep(5, 3),
  lower_corr = .4,
  upper_corr = .6,
  n = 100
)

#> # A tibble: 100 x 15
#>    block1_x1 block1_x2 block1_x3 block1_x4 block1_x5 block2_x1 block2_x2
#>        <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>
#>  1     0.354   -1.45      -0.167   -0.985     0.101     0.977     0.830 
#>  2     0.781    0.409      0.913    2.49      1.34      0.796     0.931 
#>  3    -0.810   -1.59      -1.17    -1.12     -0.673    -1.40     -1.37  
#>  4    -1.94    -1.22      -2.40    -2.06     -1.22      0.820     0.478 
#>  5    -0.252    0.769     -0.774    0.739     1.33      0.396    -0.0798
#>  6    -0.447   -0.589     -0.592   -0.0559   -0.301     0.0815    0.0135
#>  7    -2.34    -1.65      -1.20    -1.87     -0.864     1.22      0.580 
#>  8     1.17     0.551      0.894    1.77     -1.04     -0.205    -0.417 
#>  9    -0.991    0.0942    -0.101    0.797     0.0730   -0.290    -0.202 
#> 10    -1.92    -1.36      -2.62    -1.66     -1.37     -0.234     0.253 
#> # … with 90 more rows, and 8 more variables: block2_x3 <dbl>, block2_x4 <dbl>,
#> #   block2_x5 <dbl>, block3_x1 <dbl>, block3_x2 <dbl>, block3_x3 <dbl>,
#> #   block3_x4 <dbl>, block3_x5 <dbl>