Simulate correlated blocks of variables

simulate_block_data() creates a dataset of blocks of data where variables within each block are correlated. The correlation for each pair of variables is sampled uniformly from lower_corr to upper_corr, and the values of each are sampled using MASS::mvrnorm().

simulate_block_data(block_sizes, lower_corr, upper_corr, n,
  block_name = "block", sep = "_", var_name = "x")

Arguments

block_sizes
a vector of block sizes. The size of each block is the number of variables within it.

lower_corr
the lower bound of the correlation within each block

upper_corr
the upper bound of the correlation within each block

n
the number of observations or rows

block_name
description prepended to the variable to indicate the block it belongs to

sep
a character, what to separate the variable names with

var_name
the name of the variable within the block

Value

a tibble with sum(block_sizes) columns and n rows.

Examples

# create a 100 x 15 data set with 3 blocks
simulate_block_data(
  block_sizes = rep(5, 3),
  lower_corr = .4,
  upper_corr = .6,
  n = 100
)

#> # A tibble: 100 x 15
#>    block1_x1 block1_x2 block1_x3 block1_x4 block1_x5 block2_x1 block2_x2
#>        <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>
#>  1    0.576     -1.55     0.433     0.446      0.516    -0.317     0.246
#>  2    0.264      2.55     0.219     1.30       1.50     -1.56      0.128
#>  3   -0.368     -0.347   -0.759    -0.382     -1.17      0.295    -0.429
#>  4   -0.811     -0.246   -0.127     0.156      0.407    -0.250    -0.640
#>  5    1.48       1.12     0.0244   -0.970      0.549     0.440     0.382
#>  6   -0.535      0.801    1.44     -1.44       1.15     -0.558     2.14 
#>  7   -0.0756    -0.475   -1.00     -0.873     -0.636    -0.221    -0.211
#>  8    0.0104     0.372    1.30      0.692     -0.946    -0.302     0.120
#>  9   -0.933     -1.52    -0.862    -2.25      -2.60      0.137    -1.01 
#> 10   -0.772     -1.34    -0.639     0.0364    -1.11      0.185    -0.613
#> # … with 90 more rows, and 8 more variables: block2_x3 <dbl>, block2_x4 <dbl>,
#> #   block2_x5 <dbl>, block3_x1 <dbl>, block3_x2 <dbl>, block3_x3 <dbl>,
#> #   block3_x4 <dbl>, block3_x5 <dbl>