Simulate correlated blocks of variables

simulate_block_data() creates a dataset of blocks of data where variables within each block are correlated. The correlation for each pair of variables is sampled uniformly from lower_corr to upper_corr, and the values of each are sampled using MASS::mvrnorm().

simulate_block_data(
  block_sizes,
  lower_corr,
  upper_corr,
  n,
  block_name = "block",
  sep = "_",
  var_name = "x"
)

Arguments

block_sizes
a vector of block sizes. The size of each block is the number of variables within it.

lower_corr
the lower bound of the correlation within each block

upper_corr
the upper bound of the correlation within each block

n
the number of observations or rows

block_name
description prepended to the variable to indicate the block it belongs to

sep
a character, what to separate the variable names with

var_name
the name of the variable within the block

Value

a tibble with sum(block_sizes) columns and n rows.

Examples

# create a 100 x 15 data set with 3 blocks
simulate_block_data(
  block_sizes = rep(5, 3),
  lower_corr = .4,
  upper_corr = .6,
  n = 100
)
#> # A tibble: 100 × 15
#>    block1_x1 block1_x2 block1_x3 block1_x4 block1_x5 block2_x1 block2_x2
#>        <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>
#>  1  -0.696     -0.980    0.235      -0.118     1.59     0.877     -0.629
#>  2  -0.400      0.0457  -1.23       -0.921    -1.18    -0.583     -0.958
#>  3  -1.03      -0.259    0.411      -1.07      0.183   -0.568      0.971
#>  4  -1.15      -0.307    0.180       0.235    -0.814    0.244      0.483
#>  5   0.00749   -0.120   -0.00223    -1.02     -1.43     0.357     -0.386
#>  6  -1.18      -0.534    0.180      -1.31     -1.14    -1.45      -1.09 
#>  7   0.930      1.35    -0.709       0.512    -0.945   -0.266      1.04 
#>  8   0.584      1.46     1.48       -0.356    -1.19    -0.0244     0.656
#>  9  -1.60      -0.138    0.0832     -0.932     0.184    0.194      0.815
#> 10   0.955      1.54     0.648       1.80      0.263    0.795     -1.46 
#> # … with 90 more rows, and 8 more variables: block2_x3 <dbl>, block2_x4 <dbl>,
#> #   block2_x5 <dbl>, block3_x1 <dbl>, block3_x2 <dbl>, block3_x3 <dbl>,
#> #   block3_x4 <dbl>, block3_x5 <dbl>